added new error for user validation, fixed delelete for plant of user, dubble plant for one user

+++ src/database/plants/queryDeletePlantByPlantId.ts
@@ -6,10 +6,10 @@ const uri: string = process.env.MONGO_CONNECT_URL!;
 const database: string = process.env.DATABASE!;
 const client = new MongoClient(uri);
 
-export const queryDeletePlantByPlantId = async (plantId: string) => {
+export const queryDeletePlantByPlantIdAndUserId = async (plantId: string, userId:string) => {
 	try {
 		await connectDatabase();
-		await client.db(database).collection('Plants').deleteOne({ plantId: plantId });
+		await client.db(database).collection('Plants').deleteOne({ plantId: plantId, userId:userId });
 	} catch (error) {
 		return error;
 	} finally {
